### 2019-10-15 #1810 @whedon
[PRE REVIEW]: MatSurv: Survival analysis and visualization in MATLAB
**Submitting author:** @jhcreed (<a href="http://orcid.org/0000-0002-8302-2648">Jordan Creed</a>)
**Repository:** <a href="https://github.com/aebergl/MatSurv" target ="_blank">https://github.com/aebergl/MatSurv</a>
**Version:** v1.0.0
