### 2019-10-23 #1830 @cMadan
@jhcreed, can you please fix your paper.md? It looks like it's missing the date information.
Compare https://raw.githubusercontent.com/aebergl/MatSurv/master/Article/paper.md with https://raw.githubusercontent.com/adrn/gala/f6bb2532afeed3c0b0feea3037bf71a3f42b2a12/paper/paper.md
