+++ Article/paper.md
@@ -13,6 +13,7 @@ authors:
 affiliations: 
   - name: Moffitt Cancer Center, Tampa, FL
     index: 1
+    date: 14 October 2019
 bibliography: paper.bib
 ---
 
