+++ Article/paper.md
@@ -61,7 +61,7 @@ or censoring, and `GroupVar`, a vector or cell array defining the
 comparison groups (see example code below).
 
 ```
-    [p,fh,stats]=MatSurv([], [], [], 'Xstep', 4, 'Title', 'MatSurv KM-Plot');
+[p,fh,stats]=MatSurv([], [], [], 'Xstep', 4, 'Title', 'MatSurv KM-Plot');
 
 ```
 
@@ -88,110 +88,52 @@ SAS and MatSurv).
 ### R
 
 ```
-    fit <- survfit(survobj ~ RISK_CYTO, data=dat)
+fit <- survfit(survobj ~ RISK_CYTO, data=dat)
 
-    ggsurvplot(fit, risk.table=TRUE, pval=TRUE, risk.table.y.text.col=TRUE, risk.table.y.text=FALSE, break.time.by=24)
+ggsurvplot(fit,
+     risk.table=TRUE,
+     pval=TRUE,
+     risk.table.y.text.col=TRUE,
+     risk.table.y.text=FALSE,
+     break.time.by=24)
 ```
 
 ### SAS
 
 ```
-    proc lifetest data=lamlv2(where=(RISK_CYTO ^= 'N.D.')) intervals=(0 to 120 by 24) timelist = (0 to 120 by 24)  plots=survival(atrisk=0 to 120 by 24 test);
-    time OS_MONTHS*Surv(0);
-    strata RISK_CYTO/test=logrank;
-    run;
+proc lifetest data=lamlv2(where=(RISK_CYTO ^= 'N.D.')) 
+intervals=(0 to 120 by 24) timelist = (0 to 120 by 24)  
+plots=survival(atrisk=0 to 120 by 24 test);
+time OS_MONTHS*Surv(0);
+strata RISK_CYTO/test=logrank;
+run;
 ```
 
 ### MatSurv
 
 ```
-    load laml_RC_data.mat
+load laml_RC_data.mat
 
-    [p,fh,stats]=MatSurv(laml_RC_TimeVar, laml_RC_EventVar,  laml_RC_GroupVar,... 'GroupsToUse', {'Good', 'Intermediate', 'Poor'}, 'Xstep', 24);
+[p,fh,stats]=MatSurv(laml_RC_TimeVar, laml_RC_EventVar,
+  laml_RC_GroupVar,... 'GroupsToUse',
+   {'Good', 'Intermediate', 'Poor'}, 'Xstep', 24);
 ```
 
-![](figure_20181022.png)
+![Figure 1](figure_20181022.png)
 
 The results from MatSurv have been compared against both SAS and R and
 found to return similar estimates. The Chi-Sq values and p-calues for a
 long-rank test in MatSurv, SAS, and R are provided below (Table 1).
 
-<table>
-<thead>
-<tr class="header">
-<th>Data</th>
-<th>Groups</th>
-<th style="text-align: center;">MatSurv</th>
-<th style="text-align: center;">MatSurv</th>
-<th style="text-align: center;">SAS</th>
-<th style="text-align: center;">SAS</th>
-<th style="text-align: center;">Survminer</th>
-<th style="text-align: center;">Survminer</th>
-</tr>
-</thead>
-<tbody>
-<tr class="odd">
-<td></td>
-<td></td>
-<td style="text-align: center;"><strong>chi-sq</strong></td>
-<td style="text-align: center;"><strong>p</strong></td>
-<td style="text-align: center;"><strong>chi-sq</strong></td>
-<td style="text-align: center;"><strong>p</strong></td>
-<td style="text-align: center;"><strong>chi-sq</strong></td>
-<td style="text-align: center;"><strong>p</strong></td>
-</tr>
-<tr class="even">
-<td>Freireich</td>
-<td>Groups</td>
[…]
