### 2019-10-23 a7c5b83
remove double caption
