+++ Article/paper.md
@@ -119,7 +119,7 @@ load laml_RC_data.mat
    {'Good', 'Intermediate', 'Poor'}, 'Xstep', 24);
 ```
 
-![KM plots from Survminer, SAS and MatSurv](figure_20181022.png)
+![Output for Survminer (A), SAS (B) and MatSurv (C). All three produce the same logrank p-value of 4.02E-6](figure_20191023.png)
 
 The results from MatSurv have been compared against both SAS and R and
 found to return similar estimates. The Chi-Sq values and p-calues for a
