### 2019-10-28 #1830 @jhcreed
We have looked back over it and fixed the typos. Thanks!
