### 2019-11-05 #1830 @dsurujon
Thank you for this opportunity to review MatSurv. I have gone through the manuscript, documentation and example usage cases. I believe this is a very useful piece of code, and works well, with plenty of options for customization of the output figures. The authors have supplied a number of reproducible examples that are very helpful as well. However, I believe the authors can better highlight the utility of MatSurv by addressing the points I raise below. Once these are addressed, I will be happy to recommend the manuscript for publication.

General checks
==============

*Authorship*
I see another user @pjl54 has made a commit, but doesn't appear as an author. Depending on the extent of their contribution, I suggest including them as an author or at least mentioning them in the acknowledgements if it was a minor contribution.

Documentation
=============
*Statement of need*
Please include a description of what contexts survival analysis would be useful in, and what the target audience is.

*Community guidelines*
Please include this section in the documentation

Software paper
==============
*Statement of need*
In the summary section, I believe the authors can be more specific with some examples of when survival analysis would be applied, and what kinds of conclusions can be drawn from the output of MatSurv and similar software. I believe the Summary section could include more information on when and how survival analysis is used, since this might not be clear for a general audience.

*Quality of writing*
* I found the writing difficult to follow. The first half of the Summary section and the first paragraph of the Use section are mostly definitions, which could benefit a lot from being put into context, with descriptions of **why** these statistics are used, and **how** they give us meaningful information.
* It is not clear which statistics MatSurv outputs in the text. It appears that the "stats" output includes hazard ratios and 95% CIs calculated both with the log-rank and the Mantel-Haenszel approach. However, the section under "Use" implies only the log-rank results are returned.
* The authors make a point about how the log-rank test "will give slightly different results when compared to the Mantel-Haneszel or Cox regression approach, which is commonly used in R.", but then show agreement between their output and the R output in Table 1. This is confusing, but can be clarified by explaining what exactly is being reported in Table 1.
* The chi square statistic in Table 1 is not defined, and as it stands, it is not clear how this is relevant to the log-rank test

*References*
* For the mentioned survival analysis approaches (log rank test, hazard ration and Kaplan-Meier curves), it would be helpful to have references.
* The data sources (e.g. Freireich et al., 1963) should also be referenced.

Other comments
==============
* I highly recommend the authors explain some key concepts such as what an "event" and a "censored event" is. The risk table included in the output should also be introduced both in the paper and in the documentation.
* In the risk table, it is not immediately clear that the columns correspond to the same time points as in the KM curve figure. I suggest adding another row (representing Time, as in ```ggsurvplot```) as default.
* In the documentation, I suggest describing the datasets used in more detail, as it would add to why this software is important/useful
* While the input options are defined in the documentation, there is little information on the output (especially the ```stats``` output variable). I suggest describing each individual output statistic similar to what's included as comments in the ```MatSurv.m``` script
* Table 1 does not have a legened, and it appears that in the "Data" column, the last row has a typo (LMAL instead of LAML)

*Minor comments*

* In the documentation, under [Example with gene expression data](https://github.com/aebergl/MatSurv#example-with-gene-expression-data) There is a small typo in the comment in line 6: qurtile should be quartile
* Under [List of all input options](https://github.com/aebergl/MatSurv#list-of-all-input-options) ```TimeMAx``` should be ```TimeMax```
