### 2019-11-20 bb6a014
Update text for JOSS publication
