+++ MatSurv.m
@@ -170,6 +170,8 @@ function [varargout] = MatSurv(TimeVar, EventVar, GroupVar, varargin)
 % * 'XTickFontSize': Scalar describing Xtick font size change compared
 %   to base font size (Default: -2)
 %
+% * 'YLim': Vector defining the range of the Y-axis  (Default: [0 1])
+%
 % * 'YTicks': Vector defining the position of the X-tick marks
 %   (Default: [0:0.2:1])
 %
@@ -197,6 +199,10 @@ function [varargout] = MatSurv(TimeVar, EventVar, GroupVar, varargin)
 %   to base font size (Default: 0)
 %
 % Risk table plot options
+%
+% * 'RT_XAxis': A true/false value which, if true, a X-axis line is
+%   included in the risk table. (default: True)
+%
 % * 'RT_FontSize': Scalar describing Risk Table font size change compared
 %   to base font size (Default: 0)
 %
@@ -212,7 +218,7 @@ function [varargout] = MatSurv(TimeVar, EventVar, GroupVar, varargin)
 %   Y-axis (Default: True )
 %
 % * 'CensorInRT': True/False for whether number censored should be listed
-% in risk table (Default: False)
+%   in risk table (Default: False)
 %
 % * 'RTtitleAlignment': Where RT title should be aligned (Default: middle)
 %
@@ -336,7 +342,11 @@ else % Creat KM-Plot
         axh_KM = axes(fh,'Position',options.KM_position,'NextPlot','add','tag','KM-Plot');
         axh_RT = axes(fh,'Position',options.RT_position,'tag','Risk Table');
         % No axis for the Risk Table
-        axh_RT.XAxis.Visible='off';
+        if options.RT_XAxis
+            axh_RT.XAxis.Visible='on';
+        else
+            axh_RT.XAxis.Visible='off';
+        end
         axh_RT.YAxis.Visible='off';
     end
     
@@ -400,7 +410,7 @@ else % Creat KM-Plot
     
     %Fix Y-Axis
     % Limit range from 0 to 1
-    axh_KM.YLim = [0 1];
+    axh_KM.YLim = options.YLim;
     axh_KM.YTick = options.YTick;
     YMinorStep =  (options.YTick(2) - options.YTick(1) ) / (1+options.YMinorTick);
     axh_KM.YAxis.MinorTickValues = YMinorStep:YMinorStep:1;
@@ -475,7 +485,9 @@ else % Creat KM-Plot
    
     % And now to the Risk table
     if ~options.NoRiskTable
+        axh_RT.LineWidth = 1.5;
         axh_RT.XTick=axh_KM.XTick;
+        axh_RT.XAxis.FontSize=options.BaseFontSize + options.XTickFontSize;
         % Get number of samples for each time point
         RT_X = zeros(length(axh_KM.XTick),DATA.numGroups);
         RT_Xcensor = zeros(length(axh_KM.XTick),DATA.numGroups);
@@ -491,8 +503,8 @@ else % Creat KM-Plot
             end
             
         end
-%         axh_RT.YLim = [0.5 DATA.numGroups + 0.5];
-        axh_RT.YLim = [0 DATA.numGroups + 0.5];
+        % Create RT graph for text 
+        axh_RT.YLim = [0.1 DATA.numGroups + 0.1];
         axh_RT.YTick = 1:DATA.numGroups;
         linkaxes([axh_RT,axh_KM],'x')
         
@@ -526,7 +538,6 @@ else % Creat KM-Plot
         nudge_x = abs(axh_RT.XLim(2) - axh_RT.XLim(1))/100;
[…]
