### 2019-11-21 9f2b886
Update README.md
Updated options
