+++ Article/paper.bib
@@ -1,3 +1,12 @@
+@article{Kleinbaum1998,
+  author  = {David G Kleinbaum}, 
+  title   = {Survival analysis, a self‐learning text.},
+  journal = {Biometrical Journal: Journal of Mathematical Methods in Biosciences },
+  year    = 1998,
+  pages   = {107-108},
+}
+
+
 @misc{survminer,
   author = {A. Kassambara},
   title = {Survminer},
@@ -6,3 +15,45 @@
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/kassambara/survminer}}
 }
+
+% 31575572 
+@Article{morris2019,
+   Author="Morris, T. P.  and Jarvis, C. I.  and Cragg, W.  and Phillips, P. P. J.  and Choodari-Oskooei, B.  and Sydes, M. R. ",
+   Title="{{P}roposals on {K}aplan-{M}eier plots in medical research and a survey of stakeholder views: {K}{M}unicate}",
+   Journal="BMJ Open",
+   Year="2019",
+   Volume="9",
+   Number="9",
+   Pages="e030215",
+   Month="Sep"
+}
+
+
+% 23634996 
+@Article{laml,
+   Author="Ley, T. J.  and Miller, C.  and Ding, L.  and Raphael, B. J.  and Mungall, A. J.  and Robertson, A.  and Hoadley, K.  and Triche, T. J.  and Laird, P. W.  and Baty, J. D.  and Fulton, L. L.  and Fulton, R.  and Heath, S. E.  and Kalicki-Veizer, J.  and Kandoth, C.  and Klco, J. M.  and Koboldt, D. C.  and Kanchi, K. L.  and Kulkarni, S.  and Lamprecht, T. L.  and Larson, D. E.  and Lin, L.  and Lu, C.  and McLellan, M. D.  and McMichael, J. F.  and Payton, J.  and Schmidt, H.  and Spencer, D. H.  and Tomasson, M. H.  and Wallis, J. W.  and Wartman, L. D.  and Watson, M. A.  and Welch, J.  and Wendl, M. C.  and Ally, A.  and Balasundaram, M.  and Birol, I.  and Butterfield, Y.  and Chiu, R.  and Chu, A.  and Chuah, E.  and Chun, H. J.  and Corbett, R.  and Dhalla, N.  and Guin, R.  and He, A.  and Hirst, C.  and Hirst, M.  and Holt, R. A.  and Jones, S.  and Karsan, A.  and Lee, D.  and Li, H. I.  and Marra, M. A.  and Mayo, M.  and Moore, R. A.  and Mungall, K.  and Parker, J.  and Pleasance, E.  and Plettner, P.  and Schein, J.  and Stoll, D.  and Swanson, L.  and Tam, A.  and Thiessen, N.  and Varhol, R.  and Wye, N.  and Zhao, Y.  and Gabriel, S.  and Getz, G.  and Sougnez, C.  and Zou, L.  and Leiserson, M. D.  and Vandin, F.  and Wu, H. T.  and Applebaum, F.  and Baylin, S. B.  and Akbani, R.  and Broom, B. M.  and Chen, K.  and Motter, T. C.  and Nguyen, K.  and Weinstein, J. N.  and Zhang, N.  and Ferguson, M. L.  and Adams, C.  and Black, A.  and Bowen, J.  and Gastier-Foster, J.  and Grossman, T.  and Lichtenberg, T.  and Wise, L.  and Davidsen, T.  and Demchok, J. A.  and Shaw, K. R.  and Sheth, M.  and Sofia, H. J.  and Yang, L.  and Downing, J. R.  and Eley, G.  and Alonso, S.  and Ayala, B.  and Baboud, J.  and Backus, M.  and Barletta, S. P.  and
[…]
