+++ Article/paper.md
@@ -20,13 +20,27 @@ bibliography: paper.bib
 # Summary
 
 Survival analysis is a set of methods for evaluating time-to-event
-data that is widely applied across research disciplines. Commonly
+data that is widely applied across research disciplines. For example,
+it is commonly used in clinical trials to compare the effect of treatments.
+In cancer biology, it can be used to understand how low- or high-expression
+of genes affect the aggressiveness of the tumor. Time-to-event data 
+frequently include censored data points, samples where no event was observed.
+An event is, for example, death, relapse of disease, or a new metastatic tumor.
+If none of these events occur during the study period, the time to-to-event is
+unknown, we only know that no events were observed during the study time.
+The methods described below was developed for this kind of data.  For an
+in-depth introduction to survival analysis, we can recommend the book by
+Kleinbaum and David (Kleinbaum 1998). In fact, much of the code used in MatSurv is
+based on the equations given in the book. Commonly 
 reported elements of survival analysis include log-rank tests, hazard
 ratios (HR) and Kaplan-Meier (KM) curves. KM-curves are used to compare
 survival durations between two or more groups and give users a particular
 estimate of survival probability at a given time; log-rank tests are
 used to conduct statistical inference on survival durations between
-groups; and HRs provide a ratio of the hazard rates between groups.
+groups; and HRs provide a ratio of the hazard rates between groups. 
+To further improve the KM-plot, it has been suggested that the KM-plot
+always should be accomplished by a table that describe the number of
+patients that are still “at-risk” at a specific timepoint (Morris 2019). 
 MATLAB (MATLAB 2018A) currently lacks functions to easily create
 KM-plots with accompanying risk tables. Furthermore, MATLAB does not
 have a built-in log-rank test, nor is one available in any of the
@@ -38,20 +52,30 @@ multiple groups. In addition, MatSurv allows the user to easily modify
 the appearance of the created figure. The graphics were inspired by the
 `survminer` R-package (Kassambara 2018).
 
-# Use
+# MatSurv Use and Features
 
-MatSurv uses the log-rank (Mantel-Cox) test to calculate the p-value.
+MatSurv creates three different items, a KM-plot, a risk table and
+statistical results. The KM-plot shows the events and also censoring
+for the different groups and it is customizable using different input
+parameters. The risk table shows the number of patients “at-risk” for
+different time points and it is linked to the KM-plot. The statistical
+results are reported as a structure and the different values are described
+below. MatSurv uses the log-rank (Mantel-Cox) test to calculate the Chi-
+squared test statistic and corresponding p-value describing evidence against
+the null hypothesis that the curves are identical.
 Users have two options for calculating HRs: the log-rank
-or Mantel-Haneszel approach. In the log-rank approach, HR =
+or Mantel-Haenszel approach. HR can only be calculated when there are two
+groups being compared. In the log-rank approach, HR =
 (O<sub>a</sub>/E<sub>a</sub>)/(O<sub>b</sub>/E<sub>b</sub>), where
 O<sub>a</sub> & O<sub>b</sub> are the observed events in each group and
 E<sub>a</sub> & E<sub>b</sub> are the number of expected events. In the
 Mantel-Haenszel approach, HR = exp((O<sub>1</sub>-E<sub>1</sub>)/V),
 where O<sub>1</sub> is the number of observed events in a group,
[…]
