### 2019-11-21 179e0d6
Added an option so the raisk table can be placed in the KM-plot
