+++ MatSurv.m
@@ -48,7 +48,7 @@ function [varargout] = MatSurv(TimeVar, EventVar, GroupVar, varargin)
 % OTHER PARAMETERS (passed as parameter-value pairs)
 % * 'NoPlot': A true/false value which, if true, no figure is created
 %   (default: false)
-% 
+%
 % * 'NoRiskTable': A true/false value which, if true, no risk table is
 %   included in the KM-plot. (default: false)
 %
@@ -200,6 +200,9 @@ function [varargout] = MatSurv(TimeVar, EventVar, GroupVar, varargin)
 %
 % Risk table plot options
 %
+% * 'RT_KMplot': A true/false value which, if true, the risk table is placed
+%    as a part of the KM-plot. (default: False)
+%
 % * 'RT_XAxis': A true/false value which, if true, a X-axis line is
 %   included in the risk table. (default: True)
 %
@@ -322,10 +325,10 @@ if options.NoPlot
         indx_MST = find((yb <= 0.5),1);
         if ~isempty(indx_MST)
             stats.MedianSurvivalTime(i) = xb(indx_MST);
-%         else
-%             % revert to other median
-%             [~,indx_MST] = min(yb-median(yb));
-%             stats.MedianSurvivalTime(i) = xb(indx_MST);
+            %         else
+            %             % revert to other median
+            %             [~,indx_MST] = min(yb-median(yb));
+            %             stats.MedianSurvivalTime(i) = xb(indx_MST);
         end
     end
     fh = [];
@@ -336,7 +339,7 @@ else % Creat KM-Plot
     fh=figure('Name','MatSurv KM-Plot','Color','w','Tag','MatSurv KM-Plot figure');
     
     %Create Axes
-    if options.NoRiskTable
+    if options.NoRiskTable || options.RT_InKM_Plot
         axh_KM = axes(fh,'NextPlot','add','tag','KM-Plot');
     else
         axh_KM = axes(fh,'Position',options.KM_position,'NextPlot','add','tag','KM-Plot');
@@ -449,8 +452,13 @@ else % Creat KM-Plot
         axh_KM.XLim = [0 options.XLim];
     end
     
+    if options.RT_InKM_Plot
+        axh_KM.XLim(1) = axh_KM.XLim(1) - ((axh_KM.XLim(2)-  axh_KM.XLim(1))/20);
+    end
+    
     max_X = axh_KM.XLim(2);
     Nudge_X = max_X / 50;
+    
     
     if ~isempty(options.Xstep)
         axh_KM.XTick = 0:options.Xstep:max_X;
@@ -480,14 +488,12 @@ else % Creat KM-Plot
                 end
             end
         end
-        text(axh_KM,Nudge_X,0.1,txt_str,'FontSize',options.BaseFontSize + options.PvalFontSize,'tag','p-value')
+        text(axh_KM,axh_KM.XLim(1)+Nudge_X,0.1,txt_str,'FontSize',options.BaseFontSize + options.PvalFontSize,'tag','p-value')
     end
-   
+    
     % And now to the Risk table
     if ~options.NoRiskTable
-        axh_RT.LineWidth = 1.5;
-        axh_RT.XTick=axh_KM.XTick;
-        axh_RT.XAxis.FontSize=options.BaseFontSize + options.XTickFontSize;
+        
         % Get number of samples for each time point
         RT_X = zeros(length(axh_KM.XTick),DATA.numGroups);
         RT_Xcensor = zeros(length(axh_KM.XTick),DATA.numGroups);
@@ -503,10 +509,6 @@ else % Creat KM-Plot
             end
             
         end
-        % Create RT graph for text 
[…]
