### 2019-11-21 99bcba8
Added an option so the risk table can be placed in the KM-plot
