### 2019-11-21 3a440a6
Update paper.md
Updated Acknowledgements
