### 2019-11-22 adafa52
Update README.md
Updated Example 2
