+++ README.md
@@ -102,7 +102,7 @@ In the example below, we show how we can change some of the properties of the KM
 [p,fh,stats]=MatSurv([],[],[],'Xstep',4,...
 'TitleOptions',{'Color','r','Interpreter','none'},'InvHR',1,...
 'Xlim',32,'XMinorTick',3,'LineColor',[0 0 1;1 0 1],'LineStyle',{'-',':'},...
-'LineWidth',3,'CensorLineColor','k','RT_Title','Risk Table');
+'LineWidth',3,'CensorLineColor','k','RT_KMplot',1);
 
 ```
 <img src="/figures/Example_02.png" alt="MatSurv example" width="600">
