### 2019-11-22 97efbc8
Update README.md
Updated RT_KMplot in the option list
