+++ README.md
@@ -308,7 +308,7 @@ KM plot options
 
 Risk table plot options
 
-*`RT_KMplot`: A true/false value which, if true, the risk table is placed
+* `RT_KMplot`: A true/false value which, if true, the risk table is placed
   as a part of the KM-plot. (default: False)
 
 * `RT_XAxis`: A true/false value which, if true, a X-axis line is
