### 2019-12-12 aebergl/MatSurv#8 @ManuelaS
GroupOrder flag does not seem to work as expected
Hi,

I tried to reorder the groups from the LAML example to have “N.D.” at one extreme ('Poor', 'Intermediate', 'Good', 'N.D.') using the flag “GroupOrder” and I got the error below.
`
load('laml_RC_data.mat');

[p, fh, stats] = MatSurv(...
laml_RC_TimeVar,...
laml_RC_EventVar,...
laml_RC_GroupVar,...
'GroupOrder', {'Poor', 'Intermediate', 'Good', 'N.D.'})

`

Errors as

> Conversion to logical from cell is not possible.
   Error in MatSurv (line 293)
  if options.GroupOrder

It looks like the flag is expected to get a boolean rather than a cell array with the group order as in the doc string.
