### 2019-12-12 #1830 @ManuelaS
**Review**
Apologies for the delay to getting to this review.
I enjoyed reading through the manuscript and using MatSurv. MatSurv is well-written, easy to “install” and use, has good documentation and offers plenty of options to tailor Kaplan-Meier plots to the specific study needs. I believe MatSurv will be a useful addition for the MATLAB community.

**Below some suggestions:**

- Survival analysis is applied in many fields of research and some of those communities are strong MATLAB users. I believe keeping the description of survival analysis more general and referring to the clinical setting as one of the many applications would attract a broader usership.

- I believe it would be helpful to include in the paper that  simple Kaplan-Meier plots (without risk-tables) can be drawn and Cox regression models can be fitted in MATLAB with the Statistics and Machine Learning Toolbox.

- Include minimum required MATLAB version along with toolboxes (and when they are needed).
    - I ran into this issue when I first started reviewing it. The code checks that the MATLAB Version is
     at least 2016b, though it would be helpful for users to be made aware before downloading,
     installing and attempting to use the code.
    -  While no other toolbox are needed to use MatSurv with defaults, it requires Statistics and
       Machine Learning Toolbox when setting the flag `CutPoint` to ‘Quartile’ or ‘Tertile’ (because it
      uses the function `prctile`).

- The authors for mentioning that the statistical output from MatSurv is close, but does not match the output from R and SAS. I believe this section should be expanded with more details of why the results diverge and the impact not only on p-values, but also on hazard ratios and confidence intervals. Whether the results match or almost match outputs from “established” software is important from a user perspective and I believe this information should feature more prominently in the ReadMe as well as the paper.
- MatSurv works well and as advertised (I used MATLAB 2019b and the latest GitHub version of the code). Though I ran into a couple of issues while I was testing it out:
    - Flag `GroupOrder` does not seem to work as expected (issue:
      https://github.com/aebergl/MatSurv/issues/8)
    - P-value change when setting `GroupOrder` in some settings (issue:
      https://github.com/aebergl/MatSurv/issues/9)
- I did not see any unit-tests (which also be useful for contributors interested in submitting PRs).

**Minor comments:**
- The documentation for the flag `CutPoint` is missing the option `Tertile` in the ReadMe
- It would be helpful to match the color scheme and layout for the 3 subplots in the manuscript figure which compares the plotting capabilities of  MatSurv with those from SAS and R.
- A few typos I came across:
    - Typo in “suitiable” (“Why MatSurv” section , ReadMe)
    - Typo “If it is a continues variable” (documentation string in MatSurv.m)
    - Missing third person in “table that describe“ (“Summary” section in the paper)
