### 2019-12-12 aebergl/MatSurv#8 @aebergl
Thanks for finding this! There was some changes pushed that may have broken this and I should have done a more detailed testing. Fixes to come soon
