### 2019-12-12 3147fb0
Fixed SortOrder input so it can be either a numerical vector or a cell array ( Issue#8 )
