+++ MatSurv.m
@@ -21,7 +21,7 @@ function [varargout] = MatSurv(TimeVar, EventVar, GroupVar, varargin)
 %   'EventDefinition' can be used to define other types of events
 %
 % * 'GroupVar' is a vector or cell array defining the different groups.
-%   If it is a continues variable median cut will be used as a default.
+%   if 'GroupVar' is a numeric vector, median-cut will be used as a default.
 %
 % OUTPUTS:
 % * p       : log rank p-value
@@ -61,8 +61,10 @@ function [varargout] = MatSurv(TimeVar, EventVar, GroupVar, varargin)
 % * 'GroupsToUse': Cell array defining what groups to use from the GroupVar
 %   variable. Works only if GroupVar is a cell array. (default: all groups are used)
 %
-% * 'GroupOrder': A cell array defining the group order to be used in the
-%   legend. (default: Groups are sorted alphabetically)
+% * 'GroupOrder': A cell array or vector defining the group order to be used in the
+%   legend. The vector needs to have the same number of elements as groups 
+%   while the cell array does not have that requirement. 
+%   (default: Groups are sorthed by 'GroupToUse' if defined, else alphabetically)
 %
 % * 'EventDefinition': Two element cell array where the first cell defines
 %   the event and the second censored values. Example {'Dead','Alive'}
@@ -290,10 +292,20 @@ if options.FlipGroupOrder
     DATA.GROUPS = DATA.GROUPS(DATA.numGroups:-1:1);
 end
 
-if options.GroupOrder
+% Update group order based on GroupOrder
+if isvector(options.GroupOrder) && isnumeric(options.GroupOrder)
     DATA.GROUPS = DATA.GROUPS(options.GroupOrder);
+elseif iscell(options.GroupOrder)
+    indx_NewOrder = 1:DATA.numGroups;
+    GroupNames = [DATA.GROUPS(:).GroupName];
+    for i = length(options.GroupOrder):-1:1
+        indx = strcmp(options.GroupOrder(i),GroupNames(indx_NewOrder));
+        if any(indx)
+            indx_NewOrder = [indx_NewOrder(indx), indx_NewOrder(~indx)];
+        end
+    end
+    DATA.GROUPS = DATA.GROUPS(indx_NewOrder);
 end
-
 
 % Creat Survival table for plotting
 [DATA] = MatSurvCreateTable(DATA);
