### 2019-12-12 eb8e7d9
Updated description of GroupOrder option
