### 2019-12-12 990ca32
Updated description of CutPoint
