### 2019-12-13 aebergl/MatSurv#8 @aebergl
Fixed. It can now handle both a cell array and a scalar vector as input for GroupOrder.
