### 2019-12-13 eaaa884
Not dependent on prctile command anymore, only uses it if Statistics and Machine Learning Toolbox is available
