### 2019-12-13 3bd56f3
Update README.md
Included version requirement in description
