### 2019-12-15 aebergl/MatSurv#9 @aebergl
Hello,
Thanks for taking time to find this bug! It turns out that the calculations of the log rank p-value do not work when there are **two** or more groups with **no events**. I have added a error checking for the groups that warns about this condition but still displays the KM-plot. I also by default remove groups with less than two samples. I have also added a so one can easily merge Groups with a multi-level cell structure as GroupsToUse input variable.
Your example would be fixed using:
[p, fh, stats] = MatSurv( data.overall_survival_months,data.is_censored_overall_survival=='no', data.tumour_site, 'GroupsToUse', {{'appendix+others','appendix','others'}, 'proximal', 'distal', 'rectal'});
