### 2019-12-15 4a90227
Additional error checking including minimum number of samples per group and if there is more than one group with zero events. Also added support for multilevel GroupsToUse input in order to merge groups
