+++ MatSurv.m
@@ -59,11 +59,13 @@ function [varargout] = MatSurv(TimeVar, EventVar, GroupVar, varargin)
 %   cut points. (default: 'Median')
 %
 % * 'GroupsToUse': Cell array defining what groups to use from the GroupVar
-%   variable. Works only if GroupVar is a cell array. (default: all groups are used)
+%   variable. Groups can be merged using a multilevel cell structure, for example:
+%   {{'Group 1+2','Group1','Group2'},'Group3','Group4'} Group 1 & 2 will be
+%   merged and called Group 1+2 (default: all groups are used)
 %
 % * 'GroupOrder': A cell array or vector defining the group order to be used in the
-%   legend. The vector needs to have the same number of elements as groups 
-%   while the cell array does not have that requirement. 
+%   legend. The vector needs to have the same number of elements as groups
+%   while the cell array does not have that requirement.
 %   (default: Groups are sorthed by 'GroupToUse' if defined, else alphabetically)
 %
 % * 'EventDefinition': Two element cell array where the first cell defines
@@ -71,6 +73,9 @@ function [varargout] = MatSurv(TimeVar, EventVar, GroupVar, varargin)
 %
 % * 'TimeMin': Scalar defining minimum valid time point. Subjects with time
 %   values below this will be removed. (default: 0)
+%
+% * 'MinNumSamples': Scalar defining minimum number of samples for a Group
+%   Groups with less samples will be removed. (default: 0)
 %
 % * 'TimeMax': Scalar value defining right censoring time. Subjects with
 %   TimeVar > TimeMax will be set to TimeMax and considered as censored.
@@ -287,6 +292,10 @@ end
 % CreatGroups based on GroupVar and create DATA structure
 [DATA,options] = MatSurvCreateGroups(TimeVar, EventVarBin, GroupVar, options);
 
+% Check if there is enough samples in each group and if there is more than
+% one group with no events
+[DATA, options] = MatSurvCheckGroups(DATA, options);
+
 % Flip Group Ordering
 if options.FlipGroupOrder
     DATA.GROUPS = DATA.GROUPS(DATA.numGroups:-1:1);
@@ -311,8 +320,12 @@ end
 [DATA] = MatSurvCreateTable(DATA);
 
 % Do log rank test
-[p,stats] = MatSurvLogRank(DATA);
-
+if options.CalcP
+    [p,stats] = MatSurvLogRank(DATA);
+else
+    p=[];
+    stats=[];
+end
 if options.PairWiseP
     counter = 0;
     stats.ParwiseName = cell(DATA.numGroups * (DATA.numGroups - 1) / 2,1);
@@ -483,7 +496,7 @@ else % Creat KM-Plot
     axh_KM.XAxis.MinorTickValues = XMinorStep:XMinorStep:axh_KM.XTick(end);
     axh_KM.LineWidth = 1.5;
     
-    if options.DispP
+    if options.CalcP && options.DispP
         txt_str(1) = {sprintf('p = %.3g',p)};
         if options.DispHR && isfield(stats,'HR_logrank')
             if ~options.Use_HR_MH
@@ -610,7 +623,7 @@ else % Creat KM-Plot
     end
 end
 
-if options.Print
+if options.CalcP && options.Print
     fprintf('\n')
     fprintf('p = %.3g\n',stats.p_MC)
     if options.CalcHR && isfield(stats,'HR_logrank')
[…]
