### 2019-12-17 8c33526
Added Unit test
