### 2019-12-17 ef124a9
Added more tests
