+++ UnitTest/testMatSurv.m
@@ -1,6 +1,5 @@
 %  MatSurv test
 
-
 tol = 1e-8;
 
 % preconditions
@@ -14,7 +13,7 @@ catch
 end
 
 try
-[p,fh,stats] = MatSurv([],[],[],'NoPlot',false,'Print',false);
+    [p,fh,stats] = MatSurv([],[],[],'NoPlot',false,'Print',false);
 catch
     error('Could not run MatSurv')
 end
@@ -42,7 +41,7 @@ end
 try 
     load laml_RC_Results_01
 catch
-    error('Could not load laml_RC_Results data')
+    error('Could not load laml_RC_Results_01 data')
 end
 
 try
@@ -70,13 +69,143 @@ end
 try 
     load laml_RC_Results_02
 catch
-    error('Could not load laml_RC_Results data')
+    error('Could not load laml_RC_Results_02 data')
 end
 
 try
      [p,fh,stats]=MatSurv(laml_RC_TimeVar,laml_RC_EventVar, laml_RC_GroupVar,...
          'GroupsToUse', {'Good','Intermediate',{'ND+Poor','Poor','N.D.'}},...
-         'Xstep',24,'Print',false,'Xstep',12,'Print',false,'TimeMax',60);
+         'Xstep',24,'Print',false,'Xstep',12,'TimeMax',60);
+catch
+    error('Could not run MatSurv')
+end
+
+assert(strcmp(get(fh, 'type'), 'figure') & isa(fh, 'matlab.ui.Figure') );
+
+assert(abs(p - p_true) < tol)
+assert(abs(stats.p_MC - stats_true.p_MC) < tol)
+assert(abs(stats.Chi2_MC - stats_true.Chi2_MC) < tol)
+
+close(fh);
+
+%% Test 4: laml_RC_data pairwise comparisons
+try 
+    load laml_RC_data
+catch
+    error('Could not load laml_RC_data')
+end
+
+try 
+    load laml_RC_Results_03
+catch
+    error('Could not load laml_RC_Results_03 data')
+end
+
+try
+     [p,fh,stats]=MatSurv(laml_RC_TimeVar,laml_RC_EventVar, laml_RC_GroupVar,...
+         'PairwiseP',true,'Xstep',24,'Print',false,'Xstep',12,...
+         'GroupsToUse', {'Good','Intermediate','Poor'});
+catch
+    error('Could not run MatSurv')
+end
+
+assert(strcmp(get(fh, 'type'), 'figure') & isa(fh, 'matlab.ui.Figure') );
+
+assert(abs(p - p_true) < tol)
+assert(abs(stats.p_MC - stats_true.p_MC) < tol)
+assert(abs(stats.Chi2_MC - stats_true.Chi2_MC) < tol)
+
+for i=1:3
+    assert(abs(p - p_true) < tol)
+    assert(abs(stats.ParwiseStats(i).p_MC - stats_true.ParwiseStats(i).p_MC) < tol)
+    assert(abs(stats.ParwiseStats(i).Chi2_MC - stats_true.ParwiseStats(i).Chi2_MC) < tol)
+    assert(abs(stats.ParwiseStats(i).HR_logrank - stats_true.ParwiseStats(i).HR_logrank) < tol)
+    assert(abs(stats.ParwiseStats(i).HR_logrank_Inv - stats_true.ParwiseStats(i).HR_logrank_Inv) < tol)
+    assert(abs(stats.ParwiseStats(i).HR_MH - stats_true.ParwiseStats(i).HR_MH) < tol)
[…]
