+++ UnitTest/testMatSurv.m
@@ -197,9 +197,9 @@ catch
 end
 
 try 
-    load laml_HGF_gene_Results_02
+    load laml_HGF_gene_Results_03
 catch
-    error('Could not load laml_HGF_gene_Results_02 data')
+    error('Could not load laml_HGF_gene_Results_03 data')
 end
 
 try
