+++ Article/paper.md
@@ -125,6 +125,7 @@ ggsurvplot(fit,
      pval=TRUE,
      risk.table.y.text.col=TRUE,
      risk.table.y.text=FALSE,
+     palette = c("#445694", "#A23A2E", "#01665E"),
      break.time.by=24)
 ```
 
@@ -146,10 +147,11 @@ load laml_RC_data.mat
 
 [p,fh,stats]=MatSurv(laml_RC_TimeVar, laml_RC_EventVar,
   laml_RC_GroupVar,... 'GroupsToUse',
-   {'Good', 'Intermediate', 'Poor'}, 'Xstep', 24);
+   {'Good', 'Intermediate', 'Poor'}, 'Xstep', 24,…
+'LineColor',[0.2667,0.3373,0.5804;0.6353,0.2275,0.1804;0.0039,0.4000,0.3686]);
 ```
 
-![Output for Survminer (A), SAS (B) and MatSurv (C). All three produce the same log-rank p-value of 4.02E-6](figure_20191023.png)
+![Output for Survminer (A), SAS (B) and MatSurv (C). All three produce the same log-rank p-value of 4.02E-6](figure_20191226.png)
 
 The results from MatSurv have been compared against both SAS and R and
 found to return similar estimates. The Chi-Sq values and p-values for a
