### 2019-12-27 #1830 @aebergl
>We would like to start with thanking @dsurujon for his excellent work and insightful comments. Below you will find our changes and response to the individual comments in bold below.
> # General checks
> _Authorship_
> I see another user @pjl54 has made a commit, but doesn't appear as an author. Depending on the extent of their contribution, I suggest including them as an author or at least mentioning them in the acknowledgements if it was a minor contribution. **We have now added Patrick Leo to the acknowledgements**
>
> # Documentation
> _Statement of need_
> Please include a description of what contexts survival analysis would be useful in, and what the target audience is. **This have now been added**
>
> _Community guidelines_
> Please include this section in the documentation **This is now included**
>
> # Software paper
> _Statement of need_
> In the summary section, I believe the authors can be more specific with some examples of when survival analysis would be applied, and what kinds of conclusions can be drawn from the output of MatSurv and similar software. I believe the Summary section could include more information on when and how survival analysis is used, since this might not be clear for a general audience. **We have rewritten this section and also added more text**
>
> _Quality of writing_
>
> * I found the writing difficult to follow. The first half of the Summary section and the first paragraph of the Use section are mostly definitions, which could benefit a lot from being put into context, with descriptions of **why** these statistics are used, and **how** they give us meaningful information. **More text have now been added**
> * It is not clear which statistics MatSurv outputs in the text. It appears that the "stats" output includes hazard ratios and 95% CIs calculated both with the log-rank and the Mantel-Haenszel approach. However, the section under "Use" implies only the log-rank results are returned. **This have now been changed**
> * The authors make a point about how the log-rank test "will give slightly different results when compared to the Mantel-Haneszel or Cox regression approach, which is commonly used in R.", but then show agreement between their output and the R output in Table 1. This is confusing, but can be clarified by explaining what exactly is being reported in Table 1. **This has now been clarified in the article**
> * The chi square statistic in Table 1 is not defined, and as it stands, it is not clear how this is relevant to the log-rank test. **Chi square has now been explained**
>
> _References_
>
> * For the mentioned survival analysis approaches (log rank test, hazard ration and Kaplan-Meier curves), it would be helpful to have references. **More references has been added**
> * The data sources (e.g. Freireich et al., 1963) should also be referenced. **This reference have been added**
>
> # Other comments
> * I highly recommend the authors explain some key concepts such as what an "event" and a "censored event" is. The risk table included in the output should also be introduced both in the paper and in the documentation. **We have added text describing this in the paper and also in the documentation**
> * In the risk table, it is not immediately clear that the columns correspond to the same time points as in the KM curve figure. I suggest adding another row (representing Time, as in `ggsurvplot`) as default. **Great comment! We have now added this as a default option and also added an option to put the risk table as a part of the KM-plot**
> * In the documentation, I suggest describing the datasets used in more detail, as it would add to why this software is important/useful **We have now added text describing the data better**
> * While the input options are defined in the documentation, there is little information on the output (especially the `stats` output variable). I suggest describing each individual output statistic similar to what's included as comments in the `MatSurv.m` script **We have now added this section in the documentation**
> * Table 1 does not have a legened, and it appears that in the "Data" column, the last row has a typo (LMAL instead of LAML) **Fixed**
>
> _Minor comments_
>
> * In the documentation, under [Example with gene expression data](https://github.com/aebergl/MatSurv#example-with-gene-expression-data) There is a small typo in the comment in line 6: qurtile should be quartile **Fixed**
> * Under [List of all input options](https://github.com/aebergl/MatSurv#list-of-all-input-options) `TimeMAx` should be `TimeMax` **Fixed**
