### 2019-12-27 #1830 @aebergl
Thanks for the exellent review and great suggestions by @ManuelaS. Your comments have greatly improved the functionality of MatSurv.

**Review**
Apologies for the delay to getting to this review.
I enjoyed reading through the manuscript and using MatSurv. MatSurv is well-written, easy to “install” and use, has good documentation and offers plenty of options to tailor Kaplan-Meier plots to the specific study needs. I believe MatSurv will be a useful addition for the MATLAB community. **Thanks!**

**Below some suggestions:**

- Survival analysis is applied in many fields of research and some of those communities are strong MATLAB users. I believe keeping the description of survival analysis more general and referring to the clinical setting as one of the many applications would attract a broader usership. **We have added more text in the article to adress this comment**

- I believe it would be helpful to include in the paper that  simple Kaplan-Meier plots (without risk-tables) can be drawn and Cox regression models can be fitted in MATLAB with the Statistics and Machine Learning Toolbox. **The following text has been added to the article "The Statistics and Machine Learning Toolbox support Cox proportional hazards regression using the coxphfit function and KM-plots can be created using the plot or stairs functions."**

- Include minimum required MATLAB version along with toolboxes (and when they are needed).
    - I ran into this issue when I first started reviewing it. The code checks that the MATLAB Version is at least 2016b, though it would be helpful for users to be made aware before downloading, installing and attempting to use the code. **The MATLAB Release Compatibility has been added to the top of the README file and there is also a section describing MATLAB Release Compatibility**
    -  While no other toolbox are needed to use MatSurv with defaults, it requires Statistics and Machine Learning Toolbox when setting the flag `CutPoint` to ‘Quartile’ or ‘Tertile’ (because it uses the function `prctile`). **MatSurv is no longer dependent on `prctile`. If Statistics and Machine Learning Toolbox is available `prctile` will be used, if not, the percentiles will be calculated in a similar way**

- The authors for mentioning that the statistical output from MatSurv is close, but does not match the output from R and SAS. I believe this section should be expanded with more details of why the results diverge and the impact not only on p-values, but also on hazard ratios and confidence intervals. Whether the results match or almost match outputs from “established” software is important from a user perspective and I believe this information should feature more prominently in the ReadMe as well as the paper. **This has now been clarified in the article**
- MatSurv works well and as advertised (I used MATLAB 2019b and the latest GitHub version of the code). Though I ran into a couple of issues while I was testing it out:
    - Flag `GroupOrder` does not seem to work as expected (issue:
      https://github.com/aebergl/MatSurv/issues/8) **Fixed. It can now handle both a cell array and a scalar vector as input for GroupOrder.**
    - P-value change when setting `GroupOrder` in some settings (issue:
      https://github.com/aebergl/MatSurv/issues/9) **Fixed, It turns out that the calculations of the log rank p-value do not work when there are two or more groups with no events. We have added a error checking for the groups that warns about this condition but still displays the KM-plot. We have also added a so one can easily merge Groups with a multi-level cell structure as GroupsToUse input variable.**
      - I did not see any unit-tests (which also be useful for contributors interested in submitting PRs). **We have now added a test script for MatSurv**

**Minor comments:**
- The documentation for the flag `CutPoint` is missing the option `Tertile` in the ReadMe **Fixed**
- It would be helpful to match the color scheme and layout for the 3 subplots in the manuscript figure which compares the plotting capabilities of  MatSurv with those from SAS and R.  **Fixed**
- A few typos I came across:
    - Typo in “suitiable” (“Why MatSurv” section , ReadMe) **Fixed**
    - Typo “If it is a continues variable” (documentation string in MatSurv.m) **Fixed**
    - Missing third person in “table that describe“ (“Summary” section in the paper) **Fixed**

_Originally posted by @ManuelaS in https://github.com/openjournals/joss-reviews/issues/1830#issuecomment-564997236_
