### 2020-01-16 #1830 @dsurujon
It looks like the article text has been updated substantially. Could you please compile an updated article proof?
