### 2020-01-21 #1830 @ManuelaS
@aebergl & @cMadan, the changes to the ReadMe and manuscript along with the addition of unit tests and bug fixes look good to me.
Regarding the unit tests, I am guessing the "ground truth" stats included in the mat files that the test script checks against were generated by MatSurv. I wonder (no review-breaking) whether adding some tests that compare the output of MatSurv to the statistical results from an independent software (for example the survival package from R) would help in catching potential errors in the computations arising in corner cases.

P.S. It looks like a couple of references may be missing the DOI

I am happy to recommend acceptance!
