+++ Article/paper.bib
@@ -2,6 +2,7 @@
   author  = {David G Kleinbaum}, 
   title   = {Survival analysis, a self‐learning text.},
   journal = {Biometrical Journal: Journal of Mathematical Methods in Biosciences },
+  doi = {10.1007/978-1-4419-6646-9},
   year    = 1998,
   pages   = {107-108},
 }
@@ -20,6 +21,7 @@
 @Article{morris2019,
    Author="Morris, T. P.  and Jarvis, C. I.  and Cragg, W.  and Phillips, P. P. J.  and Choodari-Oskooei, B.  and Sydes, M. R. ",
    Title="{{P}roposals on {K}aplan-{M}eier plots in medical research and a survey of stakeholder views: {K}{M}unicate}",
+   doi = {10.1136/bmjopen-2019-030215},
    Journal="BMJ Open",
    Year="2019",
    Volume="9",
