### 2020-01-24 5e0fe0e
Update README.md
Spelling fixes
