### 2020-02-01 #1830 @Kevin-Mattheus-Moerman
@jhcreed below are some remaining minor points in relation to the paper. You can implement them and call `@whedon generate pdf` here to recreate the pdf. Thanks.

- [ ] Can you type out the state and country in your affiliation? E.g.
`Moffitt Cancer Center, Tampa, Florida, United States`
- [ ] In `The methods described below was developed for...`, methods is plural so use "were" i.e.:
`The methods described below were developed for`
- [ ] Change `.. always should be ...` to `.. should always be ...`
- [ ] Would it be possible to update figure 1 to a 3x1 panel? The A and B parts appear rather small now, especially the font and line width compared to the C part of that figure. If you place A, B, and C all under each other with the same larger font size and graph edge widths (and matched axis ticks?) this would be a big improvement.
