### 2020-02-12 #1830 @danielskatz
Can you reduce the size of the figures very sightly - right now, they are slightly bigger than the page (they overlap the page footer).

I'm also confused by the "(MATLAB 2019B)" in the text, which doesn't seem to match anything in the reference list.  Similarly, Freireich 1963 and Ley 2013.  Looking at the .md, this turns out to be because you are doing the references manually - please see the [JOSS example paper and bib file](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and adjust your paper to correctly refer to the bib file entries - you will need to add these three additional papers to the bib file as well.

Then again let me know when you think you are ready to proceed.
