+++ Article/paper.md
@@ -41,7 +41,7 @@ groups; and HRs provide a ratio of the hazard rates between groups.
 To further improve the KM-plot, it has been suggested that the KM-plot
 should always be accomplished by a table that describes the number of
 patients that are still “at-risk” at a specific timepoint (Morris 2019). 
-MATLAB (MATLAB 2019B) currently lacks functions to easily create
+MATLAB [as of version MATLAB 2019B] currently lacks functions to easily create
 KM-plots with accompanying risk tables. Furthermore, MATLAB does not
 have a built-in log-rank test, nor is one available in any of the
 existing toolboxes, including the Statistics and Machine Learning
@@ -151,7 +151,7 @@ load laml_RC_data.mat
 'LineColor',[0.2667,0.3373,0.5804;0.6353,0.2275,0.1804;0.0039,0.4000,0.3686]);
 ```
 
-![Output for Survminer (A), SAS (B) and MatSurv (C). All three produce the same log-rank p-value of 4.02E-6](figure_20191226.png =250x)
+![Output for Survminer (A), SAS (B) and MatSurv (C). All three produce the same log-rank p-value of 4.02E-6](figure_20191226.png)
 
 The results from MatSurv have been compared against both SAS and R and
 found to return similar estimates. The Chi-Sq values and p-values for a
