### 2020-02-12 #1830 @jhcreed
The figure has been reduced so that it now fits on one page without running into the footer. We have also changed "(MATLAB 2019B)" to "[as of version MATLAB 2019B]" to clarify that we are specifying a version of software and not a reference. All of the references have been updated in the .md and .bib.  We are ready to try to proceed again.
