### 2020-02-12 #1830 @danielskatz
It still seems like Matlab 2019b should be a reference that should be in the .bib file and in the .md file, otherwise, the "MATLAB 2019B" doesn't refer to anything that the reader can find.
