### 2020-02-12 #1830 @danielskatz
the first reference also needs to be fix for the first author's name
