+++ Article/paper.bib
@@ -58,4 +58,12 @@
     eprint = {http://www.bloodjournal.org/blood/article-pdf/21/6/699/571437/699.pdf},
 }
 
+@book{MATLAB:2019, 
+year = {2019}, 
+author = {MATLAB}, 
+title = {version 9.7.0.1296695 (R2019b)}, 
+publisher = {The MathWorks Inc.}, 
+address = {Natick, Massachusetts} 
+}
+
 
