+++ Article/paper.bib
@@ -44,7 +44,7 @@
 }
 
 @article{freireich,
-    author = {Acute Leukemia Group B and Freireich, Emil J and Gehan, Edmund and Frei, Emil, III and Schroeder, Leslie R. and Wolman, Irving J. and Anbari, Rachad and Burgert, E. Omar and Mills, Stephen D. and Pinkel, Donald and Selawry, Oleg S. and Moon, John H. and Gendel, B. R. and Spurr, Charles L. and Storrs, Robert and Haurani, Farid and Hoogstraten, Barth and Lee, Stanley},
+    author = {Freireich, Emil J and Gehan, Edmund and Frei, Emil, III and Schroeder, Leslie R. and Wolman, Irving J. and Anbari, Rachad and Burgert, E. Omar and Mills, Stephen D. and Pinkel, Donald and Selawry, Oleg S. and Moon, John H. and Gendel, B. R. and Spurr, Charles L. and Storrs, Robert and Haurani, Farid and Hoogstraten, Barth and Lee, Stanley and Acute Leukemia Group B},
     title = "{The Effect of 6-Mercaptopurine on the Duration of Steroid-induced Remissions in Acute Leukemia: A Model for Evaluation of Other Potentially Useful Therapy}",
     journal = {Blood},
     volume = {21},
