### 2020-02-13 1ecd137
Added JOSS publication
