### 2020-02-13 #1830 @arfon
> 👋 @openjournals/dev & @arfon - note that reference 2 in the XML has the same URL problem we've seen before the github URLs - Once this finishes processing, I will leave it open for you to manually fix this problem

Thanks. I've fixed this now.
